requirements.txt
# dependencies

discord.py==1.7.3
aiofile==3.7.4